signer interface matched with other plugins
