+++ src/plugins/SessionSigGenerator/ISessionSigGenerator.ts
